[WIP] ansible-test: add vmware-integration
VMware tests depend on a list of hosts and they assume the hosts are in
an inventory. We can extend Ansible internal inventory on the fly with
`add_host` but this increase the complexity and slow down the tests.

With this patch, we can use a inventory file dedicated to vmware.
The inventory file is stored in `test/inventory/inventory.vmware` and
will be generated from the `test/inventory/inventory.vmware.template`
file, like `network-integration` and `windows-integration`.

This patch has still the WIP status because:

- the `inventory.vmware` file is not autogenerated yet

```shell
ansible-test vmware-integration -vvv --python 3.7 vcenter_folder
```